+++ packages/smarthr-ui/src/components/AppHeader/components/common/AppLauncherSortDropdown.tsx
@@ -15,8 +15,9 @@ import { Button } from '../../../Button'
 import { Dropdown, DropdownContent, DropdownTrigger } from '../../../Dropdown'
 import { FaCaretDownIcon, FaCheckIcon } from '../../../Icon'
 import { useTranslate } from '../../hooks/useTranslate'
-import { Launcher } from '../../types'
 import { Translate } from '../common/Translate'
+
+import type { Launcher } from '../../types'
 
 const classNameGenerator = tv({
   slots: {