+++ packages/smarthr-ui/src/components/AppHeader/components/mobile/NavigationItem.tsx
@@ -1,13 +1,14 @@
-import React, { FC, useContext } from 'react'
+import React, { type FC, useContext } from 'react'
 import { tv } from 'tailwind-variants'
 
-import { Navigation } from '../../types'
 import { isChildNavigation } from '../../utils'
 import { CommonButton, commonButtonClassNameGenerator } from '../common/CommonButton'
 import { Translate } from '../common/Translate'
 
 import { MenuButton } from './MenuButton'
 import { NavigationContext } from './NavigationContext'
+
+import type { Navigation } from '../../types'
 
 const navigationItem = tv({
   base: ['[&&]:shr-px-0.5'],